Further centralisation f logic

+++ Aphid.h
@@ -6,11 +6,11 @@ class Aphid : public Creature, public CreatureInteractor
 {
 protected:
 	void suicide();
-	float eat();
-	float getDefaultNutritionalValue();
+	double eat();
+	double getDefaultNutritionalValue();
 	void interactImpl(Creature * creature) { creature->interactWith(*this); }
 	int getDirection();
-	float getMoveProbability();
+	double getMoveProbability();
 public:
 	void interactWith(CreatureInteractor & creature);
 	void interact(Aphid & creature);
